Review, move flow tests to plugin package, update templatess
review fixes

update shift state templates

update shift state templates

remove some redundant classes

move flow tests
