+++ ephios/plugins/baseshiftstructures/templates/baseshiftstructures/fragment_groups_common_state.html
@@ -1,0 +1,31 @@
+{% load i18n %}
+
+<div id="structuregroups-popover-title-{{ shift.id }}" class="d-none">
+
+</div>
+<div id="structuregroups-popover-content-{{ shift.id }}" class="d-none">
+    <i class="far fa-circle text-body-secondary"></i>
+    {% translate "participants needed" %}
+    <br>
+    <i class="far fa-check-circle text-body-secondary"></i>
+    {% translate "enough participants" %}
+    <br/>
+    <i class="fa fa-check-circle text-body-secondary"></i>
+    {% translate "full" %}
+</div>
+
+{% block groups %}
+{% endblock %}
+
+{% with stats=shift.get_signup_stats %}
+    <span class="pe-1">
+        <strong>{{ stats.confirmed_count }}{% if stats.max_count %}/{{ stats.max_count }}{% endif %}</strong>
+        {% translate "Total" context "participation count" %}
+    </span>
+{% endwith %}
+
+{% if disposition_url %}
+    <div class="mt-1">
+        <a href="{{ disposition_url }}" class="btn btn-sm btn-info">{% translate "Disposition" %}</a>
+    </div>
+{% endif %}
