+++ tests/plugins/basesignupflows/test_common.py
@@ -1,7 +1,7 @@
 from django.urls import reverse
 
 from ephios.core.models import AbstractParticipation, LocalParticipation
-from ephios.core.signup.flow.builtin.participant import InstantConfirmSignupFlow
+from ephios.plugins.basesignupflows.flow.participant import InstantConfirmSignupFlow
 
 
 def test_getting_dispatched_state_is_overwritten_by_participant_signup(
